# docs/openapi/phase2-cloud-auth.openapi.yaml
openapi: 3.1.0
info:
  title: Hacom Cloud Phase 3 API Contract
  version: 3.0.0-process1-search
  description: |
    Phase 3 endpoint contract preserving the frozen Phase 2 Process 1 HTTP
    authentication boundary. The public gateway exposes /cloud-api and strips
    that prefix before proxying to Cloud API. Details remain in docs/api.md.
servers:
  - url: /cloud-api
    description: Same-origin Chat gateway
security:
  - hacomAccessToken: []
paths:
  /health/live:
    get:
      summary: Check whether the API process is alive
      security: []
      operationId: cloudLiveness
      responses:
        "200":
          description: API process is alive
  /health/ready:
    get:
      summary: Check whether required dependencies are ready
      security: []
      operationId: cloudReadiness
      responses:
        "200":
          description: Required Cloud dependencies are ready
        "503":
          description: A required dependency is unavailable
  /api/v1/cloud/quota:
    get:
      summary: Get the authenticated owner's quota breakdown
      operationId: getCloudQuota
      responses:
        "200":
          description: Quota owned by the authenticated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quota"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/cloud/items:
    get:
      summary: List active Cloud items
      operationId: listCloudItems
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/ItemTypeFilter"
        - $ref: "#/components/parameters/CreatedFrom"
        - $ref: "#/components/parameters/CreatedTo"
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        "200":
          description: Items scoped to the authenticated subject
        "400":
          description: Invalid, repeated, unsupported, or cursor-mismatched filter
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/cloud/items/{itemId}:
    get:
      summary: Get an owned active Cloud item
      operationId: getCloudItem
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: Item owned by the authenticated subject
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Item is missing or belongs to another subject
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      summary: Permanently delete an owned active or Trash item
      operationId: permanentlyDeleteCloudItem
      parameters:
        - $ref: "#/components/parameters/ItemId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Logical delete committed; async says whether binary cleanup remains
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermanentDelete"
        "404":
          description: Item is missing or belongs to another subject
        "409":
          description: Invalid state, idempotency conflict, or delete already pending
  /api/v1/cloud/items/{itemId}/trash:
    post:
      summary: Move an owned ready item to Trash
      operationId: moveCloudItemToTrash
      parameters:
        - $ref: "#/components/parameters/ItemId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Item is in Trash for exactly 24 hours
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrashLifecycle"
        "404":
          description: Item is missing or belongs to another subject
        "409":
          description: Invalid state or idempotency conflict
  /api/v1/cloud/items/{itemId}/restore:
    post:
      summary: Restore an owned Trash item before expiry
      operationId: restoreCloudTrashItem
      parameters:
        - $ref: "#/components/parameters/ItemId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Trash item restored before purgeAfter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrashLifecycle"
        "404":
          description: Item is missing or belongs to another subject
        "409":
          description: Invalid state, expired restore, or idempotency conflict
  /api/v1/cloud/trash:
    get:
      summary: List the authenticated owner's Trash items
      operationId: listCloudTrash
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/ItemTypeFilter"
        - $ref: "#/components/parameters/CreatedFrom"
        - $ref: "#/components/parameters/CreatedTo"
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        "200":
          description: Trash items scoped to the authenticated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrashPage"
        "400":
          description: Invalid, repeated, unsupported, or cursor-mismatched filter
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/cloud/texts:
    post:
      summary: Create a text item
      operationId: createCloudText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [content]
              properties:
                content:
                  type: string
      responses:
        "201":
          description: Text created for the authenticated subject
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/cloud/links:
    post:
      summary: Create a link item
      operationId: createCloudLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                title:
                  type: string
      responses:
        "201":
          description: Link created for the authenticated subject
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/cloud/uploads:
    post:
      summary: Initiate a direct-to-object-storage upload
      operationId: initiateCloudUpload
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Existing idempotent upload reservation
        "201":
          description: Upload reservation created
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
components:
  securitySchemes:
    hacomAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access JWT issued by chat-auth-service. Verifiers pin RS256 or ES256,
        resolve kid via JWKS, verify iss/aud/exp/typ=access, require
        sub/sid/jti/iat, and enforce auth-owned session/account state.
  parameters:
    SearchQuery:
      in: query
      name: q
      description: Search title, text content, link URL, or original file name
      schema:
        type: string
        minLength: 3
        maxLength: 200
    ItemTypeFilter:
      in: query
      name: type
      schema:
        type: string
        enum: [text, link, file, image, video, audio]
    CreatedFrom:
      in: query
      name: from
      description: Inclusive created-at lower bound
      schema:
        type: string
        format: date-time
    CreatedTo:
      in: query
      name: to
      description: Inclusive created-at upper bound
      schema:
        type: string
        format: date-time
    ItemId:
      in: path
      name: itemId
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
  headers:
    RequestId:
      description: Correlation identifier generated by the trusted gateway or API
      schema:
        type: string
        maxLength: 128
  responses:
    Unauthorized:
      description: Missing, malformed, invalid, expired, wrong-type, or revoked access token
      headers:
        WWW-Authenticate:
          schema:
            type: string
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Authenticated account is not active or lacks the required capability
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AuthUnavailable:
      description: Auth session/account authority is unavailable; protected writes fail closed
      headers:
        Retry-After:
          schema:
            type: integer
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Quota:
      type: object
      additionalProperties: false
      required: [limitBytes, usedBytes, activeBytes, trashBytes, reservedBytes, availableBytes, updatedAt]
      properties:
        limitBytes: { type: integer, format: int64 }
        usedBytes: { type: integer, format: int64 }
        activeBytes: { type: integer, format: int64 }
        trashBytes: { type: integer, format: int64 }
        reservedBytes: { type: integer, format: int64 }
        availableBytes: { type: integer, format: int64 }
        updatedAt: { type: string, format: date-time }
    TrashLifecycle:
      type: object
      additionalProperties: false
      required: [itemId, status, applied]
      properties:
        itemId: { type: string, format: uuid }
        status: { type: string, enum: [ready, trashed] }
        deletedAt: { type: string, format: date-time }
        purgeAfter: { type: string, format: date-time }
        applied: { type: boolean }
    PermanentDelete:
      type: object
      additionalProperties: false
      required: [itemId, status, async]
      properties:
        itemId: { type: string, format: uuid }
        status: { type: string, enum: [deleted, delete_pending] }
        async: { type: boolean }
    TrashItem:
      type: object
      additionalProperties: false
      required: [id, type, status, sizeBytes, deletedAt, purgeAfter, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [text, link, file, image, video, audio] }
        status: { type: string, const: trashed }
        title: { type: string }
        content: { type: string }
        url: { type: string, format: uri }
        sizeBytes: { type: integer, format: int64 }
        deletedAt: { type: string, format: date-time }
        purgeAfter: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TrashPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TrashItem"
        nextCursor: { type: string }
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - AUTH_REQUIRED
                - INVALID_ACCESS_TOKEN
                - SESSION_REVOKED
                - ACCOUNT_NOT_ACTIVE
                - AUTH_AUTHORITY_UNAVAILABLE
                - ITEM_NOT_FOUND
                - INVALID_TRASH_REQUEST
                - INVALID_ITEM_STATE
                - RESTORE_EXPIRED
                - DELETE_PENDING
                - IDEMPOTENCY_CONFLICT
            message:
              type: string